Moved AwgScreen function
